Initial Implementation of Search using Haystack
This is an initial implementation although functional. Currently the template for search is duplicated taking into account the use of search query sets by haystack instead of query sets. If we could detect this inside the template this could be avoided. The parameter name query was changed to q for interoperability with Haystack. This parameter is hard coded in haystack and interacts between class view and form.

Bug: Currently the search shows articles that the user cant read. This can be corrected in the override to the method __call__. An efficient way of doing it must be used.
Another option can be rewriting the form, since in Haystack the search is actually done by the form!. SearchForm.form returns a search query set captured by __call__ in the view class that then is used to create the response by create_response inside the view.