Correct get_query_set compat for Django 1.5
Without this fix, RelatedManagers (among other things) break, and start
returning all objects when combined with manager methods.  The test checks
this.

The fix is using the pattern in the release notes:

https://docs.djangoproject.com/en/dev/releases/1.6/#get-query-set-and-similar-methods-renamed-to-get-queryset

(which I've updated recently)

There are other tests which are fixed by this patch, including an
infinite recursion in ArticleViewViewTests.test_article_list_update caused
by this issue (because root.children.active() included root)