Increase slug size from 50 to 255.
In edx-platform, default course wiki article's slug is automatically set
to "{org}.{number}.{run}", which can be more than 50 characters long.

The wiki breaks if the automatically assigned slug is longer than 50
characters.

We usually use 255 character size for columns that store course_key in
edx-platform, so I believe it makes sense to use a limit of 255 for wiki
slugs, too.