Use a bootstrapped dialog when inserting images.
The current image insert used two prompt() to get the image values. This
did not allow to set the image size and did not check/help the user when
getting the alignment. Moreover, the alignment value was not checked.
Instead use one dialog with drop-down lists for alignment and size,
and create always a correct tag.