Add context to the 'Summary' string
Currently the 'Summary' string in django-wiki collides with the same
string in django-admin. It results in django-admin's version being used
if 'django.contrib.admin' was listed earlier in INSTALLED_APPS.

At least Russian translation of django-admin's 'Summary' string is quite
bad. It may be the case for other languages too.

I didn't find a clean way to make a fallback to the previous
translations. It means that 'Summary' won't be translated until new
translation are added in transifex and exported to the django-wiki
repository.