Added new plugin editsection.
The plugin allows to edit single sections. For that a markdown
extension adds [edit] links to all sections till a configurable
depth. By clicking these links only the selected section, and all
it's subsections, is edited. For the core editing functionality
the standard Edit class is used.

Detecting, as far as possible, SeText headers which start directly
after an other block element, i.e. without an empty line in between,
is done via: Generate during parsing of the raw markdown text only a
candidate list of possible headers, and verify that list after that
by using a Treeprocessor which works on the headers which were
really generated by python-markdown.