Add Redlinks plugin
Checks each `<a>` tag in the tree to see if the `href` points to a wiki
article that does not exist; if so, the CSS class wiki-broken colors the
link red (#ba0000, same color as Wikipedia uses). Otherwise, one of the
CSS classes wiki-internal or wiki-external is used depending on whether
the link is a relative link to another wiki article.