Try a workaround for py36+django31
I pass my links as list to the template, not as query result. And in one
particular tox setup, Paginator seems to have no wrapper around its object list,
so paginator.object_list.count is the Python builtin list method for counting
the frequency of an object in a list, instead of the attribute that gives the
length of the object. With `Paginator.count`, maybe the object's .count() call
fails gracefully and it will fall back to .__len__()? I'll try that.