Add support for use pyproject.toml (#1227)
Introducing hatch! A new management tool for Python Packaging that supports the pyproject and includes more goodies to the development process. 

Here's a list that summarize what we changed thanks to hatch and the pyproject file:

- Drop `setup.py`, `pytest.ini`, `Manifest.in`, and `setup.cfg` and use `pyproject.toml`.
- Drop `tox.ini` configuration and use hatch test matrix.
- Drop `Makefile` commands and use hatch scripts.
- Migrate `pytest` and `flake8` configs to the right place.
- Remove `src/wiki/core/version.py` in favor of hatch version management.
- Update the documentation.

This commit will not be possible without the further review, testing, and comments by:

Co-authored-by: Ofek Lev <ofekmeister@gmail.com>
Co-authored-by: Benjamin Balder Bach <benjamin@overtag.dk>