Rework editsection plugin (fixes #1246)
The main problem of editsection is to find the correct location of the
section to edit in the source. Previously, that was done with custom,
regex-based parsing, but that fails when code blocks contain lines which
look like a header (for example, python comments).

This patch uses another approach to find the correct sections. It uses
extended versions of the BlockParsers responsible for creating headers,
which annotate those headers with the consumed source block. Since we
know now which input was responsible for generating the header, we can
easily locate it in the markdown source.