Pass current revision to Editor.get_widget()/get_admin_widget()
Sometimes it is useful to have the current revision in the editor
widget instance, for example to create preview and upload links,
autocomplete, ...

<textarea name="content"
          data-upload-url='/wiki/my-page/_plugin/attachments/...'>

Previously, EditorBase had an instance attribute "instance", but that
doesn't make sense (there is only one global instance of Editor) and
wasn't ever filled anyway. This patch removes that, renames the
"instance" parameter in get_widget()/get_admin_widget() to "revision"
(to make clear what this parameter is) and fills it appropriately.