Fix XSS issue while browsing attachment history
It looks like the safe tag was trying to be applied to just the default, but it is being applied to the result.
If there is any javascript in the description then the safe tag is applied to it.
Then when browsing the attachments history the script is then active.